Externed htim2 and moved the microsecond functions into their own file
