+++ Ganesha_II_V2/Core/Inc/us_timer.h
@@ -1,0 +1,14 @@
+/*
+ * us_timer.h
+ *
+ *  Created on: Apr 27, 2026
+ *      Author: Dhruv Shah
+ */
+
+#ifndef INC_US_TIMER_H_
+#define INC_US_TIMER_H_
+
+uint32_t get_time_us();
+void delay_us(uint32_t us);
+
+#endif /* INC_US_TIMER_H_ */
