+++ Ganesha_II_V2/Core/Src/main.c
@@ -170,16 +170,6 @@ void HAL_GPIO_EXTI_Callback(uint16_t pin) {
 	__HAL_GPIO_EXTI_CLEAR_FLAG(pin);
 }
 
-uint32_t get_time_us() {
-	return __HAL_TIM_GET_COUNTER(&htim2);
-}
-
-void delay_us(uint32_t us) {
-	uint32_t start_time_us = get_time_us();
-
-	while (get_time_us() - start_time_us < us) {}
-}
-
 /* USER CODE END 0 */
 
 /**
